refactor: use standard clarity and stacks-common lib
Instead of copying these libraries from stacks-blockchain, this
refactoring just pulls in the crate. This allows us to avoid regularly
merging in changes to keep them in sync.

The side-effect of this change is that the subnet Clarity can no longer
support custom built-in methods for withdrawals. Instead, a new L2 boot
contract, `subnet.clar`, defines functions for withdrawing from the L2
and uses print events instead of new custom events.