fix: updates to handle new events for withdrawals
Instead of a custom event, withdrawals now use standard contract events,
from a print in the subnet contract. The observer and the tests are
updated to utilize this new event type. From the code, withdrawals now
use a `contract-call?` to the subnet contract instead of the builtin
functions.

The subnet is also modified to boot in epoch 2.1.

Resolves: #190