feat: return `block-height` from `withdraw?` funcs
`stx-wtihdraw?`, `ft-withdraw?` and `nft-withdraw?` do not return any
meaningful value on success. When attempting to make a withdrawal, users
will need the block height in which the withdrawal occured in order to
perform the second stage of the withdraw (on the L1). This changeset
makes these withdraw functions return the block height for convenience.