feat: add `RegisterAsset` op
Track the asset registration from the L1 (`register-new-(n)ft-contract`)
and trigger a corresponding call in the L2 subnet boot contract. This
call also triggers an L2 event for observers. Additionally, the
registered contracts are recorded in the contract, allowing calls to
`(n)ft-withdraw?` to error out if the asset is not registered.