feat: add block height to `commit-block` event
The event generated by a `commit-block` (on the L1) now includes the
block height for the block that is being committed, in the
`block-height` key of the event tuple. The target burnchain block height
now uses the key `target-burn-block-height`. This is useful for
debugging, making it easy to match the L1 transactions with the L2
blocks. Note that this field is not checked in the tests since it is
purely for convenience and could be left out of some implementations.