feat: add `target-chain-height` to `commit-block`
This allows us to check if the target block is anywhere in the past, not
just the last block. That means if the block commit misses its target
block, but lands in a different block that is still in the same fork as
the target block, then the block commit can be successful.